Moves more packet files
